requirements.txt
absl-py==1.2.0
asgiref==3.5.2
astunparse==1.6.3
beautifulsoup4==4.11.1
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cymem==2.0.6
Cython==0.29.28
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
elasticsearch==7.17.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
filelock==3.7.1
flatbuffers==2.0
gast==0.5.3
gdown==4.5.1
gensim==4.2.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
inexactsearch==1.0.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
langcodes==3.3.0
libclang==14.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
murmurhash==1.0.7
nltk==3.7
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pathy==0.6.2
Pillow==9.2.0
preshed==3.0.6
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.9
PySocks==1.7.1
pyspellchecker==0.6.3
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.8.1
sentence-transformers==2.2.2
sentencepiece==0.1.96
silpa-common==0.3
six==1.16.0
smart-open==5.2.1
soundex==1.1.3
soupsieve==2.3.2.post1
spacy==3.4.0
spacy-legacy==3.0.9
spacy-loggers==1.0.3
spellchecker==0.4
sqlparse==0.4.2
srsly==2.4.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
thinc==8.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
transformers==4.20.1
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
wasabi==0.9.1
Werkzeug==2.2.0
wrapt==1.14.1
